added support for workflows in Jira DC
