+++ packages/jira-adapter/src/filters/workflow/types.ts
@@ -163,8 +163,6 @@ const statusSchema = Joi.object({
 }).unknown(true)
 
 export type Workflow = {
-  transitionIds?: Record<string, string>
-  stepIds?: Record<string, string>
   id?: Id
   entityId?: string
   name?: string
@@ -173,8 +171,6 @@ export type Workflow = {
 }
 
 export const workflowSchema = Joi.object({
-  transitionIds: Joi.object().pattern(Joi.string(), Joi.string()).optional(),
-  stepIds: Joi.object().pattern(Joi.string(), Joi.string()).optional(),
   id: idSchema.optional(),
   entityId: Joi.string().optional(),
   name: Joi.string().optional(),
